requirements.txt
openai-whisper
ffmpeg-python
numpy 
faster-whisper